no-mistakes(review): fix e2e org seed auth, dedupe org parsing, guard seed script
