+++ front-end/e2e/helpers/seeds.ts
@@ -33,15 +33,43 @@ function marketNameToSlug(name: string): string {
 }
 
 /**
+ * Log in via the API so the request context carries a Flask session cookie.
+ * Returns the user UUID.
+ */
+export async function loginViaApi(
+  request: APIRequestContext,
+  baseURL: string,
+  email: string,
+  password: string,
+): Promise<string> {
+  const loginRes = await request.post(`${baseURL}/login`, {
+    data: { email, password },
+    headers: { 'Content-Type': 'application/json' },
+  });
+  if (!loginRes.ok()) {
+    throw new Error(`Login failed: ${loginRes.status()} ${await loginRes.text()}`);
+  }
+  const loginBody = await loginRes.json() as {
+    user_data: { id: string; email: string };
+  };
+  return loginBody.user_data.id;
+}
+
+/**
  * Helper: ensure the test user has at least one organization.
- * Creates one named 'E2E Test Org' if none exist.
+ * Logs in first, since the /organizations endpoints are login-protected and
+ * the caller may pass a request context that has never authenticated.
+ * Creates an organization named 'E2E Test Org' if none exist.
  * Returns the organization ID.
  */
 export async function ensureTestOrg(
   request: APIRequestContext,
   baseURL: string,
   email: string,
+  password: string,
 ): Promise<string> {
+  await loginViaApi(request, baseURL, email, password);
+
   const orgsRes = await request.get(`${baseURL}/organizations`, {
     headers: { 'X-Owner-Email': email },
   });
@@ -87,19 +115,9 @@ export async function seedMarketWithVendors(
   password: string,
 ): Promise<SeedResult> {
   // Step 1: Login to get a session cookie and the user UUID
-  const loginRes = await request.post(`${baseURL}/login`, {
-    data: { email, password },
-    headers: { 'Content-Type': 'application/json' },
-  });
-  if (!loginRes.ok()) {
-    throw new Error(`Login failed: ${loginRes.status()} ${await loginRes.text()}`);
-  }
-  const loginBody = await loginRes.json() as {
-    user_data: { id: string; email: string };
-  };
-  const userId = loginBody.user_data.id;
+  const userId = await loginViaApi(request, baseURL, email, password);
 
-  const orgId = await ensureTestOrg(request, baseURL, email);
+  const orgId = await ensureTestOrg(request, baseURL, email, password);
 
   // Step 2: Create a market (user must own it)
   const marketName = `E2E Market ${Date.now()}`;
@@ -169,19 +187,9 @@ export async function seedPublishedMarketWithAssignments(
   password: string,
 ): Promise<PublishedSeedResult> {
   // Step 1: Login
-  const loginRes = await request.post(`${baseURL}/login`, {
-    data: { email, password },
-    headers: { 'Content-Type': 'application/json' },
-  });
-  if (!loginRes.ok()) {
-    throw new Error(`Login failed: ${loginRes.status()} ${await loginRes.text()}`);
-  }
-  const loginBody = await loginRes.json() as {
-    user_data: { id: string; email: string };
-  };
-  const userId = loginBody.user_data.id;
+  const userId = await loginViaApi(request, baseURL, email, password);
 
-  const orgId = await ensureTestOrg(request, baseURL, email);
+  const orgId = await ensureTestOrg(request, baseURL, email, password);
 
   // Step 2: Create a market
   const marketName = `E2E Published ${Date.now()}`;
@@ -316,5 +324,5 @@ export async function seedPublishedMarketWithAssignments(
     throw new Error(`Assignment store failed: ${storeRes.status()} ${await storeRes.text()}`);
   }
 
-  return { marketId, userId, marketName, marketSlug };
+  return { marketId, userId, marketName, orgId, marketSlug };
 }
