+++ front-end/e2e/market-pipeline.spec.ts
@@ -8,7 +8,7 @@ const CSV_PATH = path.resolve(__dirname, 'fixtures', 'test-vendors.csv');
 
 test.describe('Market pipeline E2E', () => {
   test.beforeAll(async ({ request }) => {
-    await ensureTestOrg(request, BACKEND_URL, TEST_USER.email);
+    await ensureTestOrg(request, BACKEND_URL, TEST_USER.email, TEST_USER.password);
   });
 
   /**
