+++ front-end/src/components/elements/ElementOrgSelect.vue
@@ -1,29 +1,9 @@
 <script setup lang="ts">
 import { ref, onMounted } from 'vue';
-import {
-    type Organization,
-    type OrganizationRoleType,
-    type ThemeObject,
-} from '@/assets/types/datatypes';
-import { api, getApiErrorMessage } from '@/utils/api';
-
-type RawOrganization = {
-    id: string;
-    name: string;
-    owner: string;
-    admins?: string[];
-    members?: string[];
-    markets?: string[];
-    ownerEmail?: string;
-    owner_email?: string;
-    adminEmails?: string[];
-    admin_emails?: string[];
-    memberEmails?: string[];
-    member_emails?: string[];
-    theme?: ThemeObject;
-    userRole?: OrganizationRoleType;
-    user_role?: OrganizationRoleType;
-};
+import { RouterLink } from 'vue-router';
+import { type Organization } from '@/assets/types/datatypes';
+import { getApiErrorMessage } from '@/utils/api';
+import { fetchOrganizations } from '@/utils/organizations';
 
 const model = defineModel<string>({ required: true });
 
@@ -31,29 +11,11 @@ const organizations = ref<Organization[]>([]);
 const loading = ref(true);
 const errorMessage = ref('');
 
-function parseOrgFromApi(org: RawOrganization): Organization {
-    const userRole = org.userRole ?? org.user_role;
-    return {
-        id: org.id,
-        name: org.name,
-        owner: org.owner,
-        admins: org.admins || [],
-        members: org.members || [],
-        markets: org.markets || [],
-        ownerEmail: org.ownerEmail ?? org.owner_email,
-        adminEmails: org.adminEmails ?? org.admin_emails,
-        memberEmails: org.memberEmails ?? org.member_emails,
-        theme: org.theme,
-        userRole: userRole ?? undefined,
-    };
-}
-
-async function fetchOrganizations() {
+async function loadOrganizations() {
     loading.value = true;
     errorMessage.value = '';
     try {
-        const response = await api.get('/organizations');
-        organizations.value = (response.data.organizations || []).map(parseOrgFromApi);
+        organizations.value = await fetchOrganizations();
     } catch (err) {
         errorMessage.value = getApiErrorMessage(err, 'Failed to load organizations');
         organizations.value = [];
@@ -63,7 +25,7 @@ async function fetchOrganizations() {
 }
 
 onMounted(() => {
-    fetchOrganizations();
+    loadOrganizations();
 });
 </script>
 
@@ -88,7 +50,7 @@ onMounted(() => {
         <p v-else-if="errorMessage" class="org-select-error">{{ errorMessage }}</p>
         <p v-else-if="organizations.length === 0" class="org-select-hint">
             No organizations available.
-            <a href="/organizations" class="org-select-link">Create an organization</a>
+            <RouterLink to="/organizations" class="org-select-link">Create an organization</RouterLink>
         </p>
     </div>
 </template>
